Start runtime init thread working.
