+++ src/runtime/monitor/src/lib.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use tracing::{debug, info, trace, Level};
+use tracing::{debug, info, trace, warn, Level};
 use tracing_subscriber::FmtSubscriber;
 
 use crate::runtime::init_actions;
@@ -29,14 +29,11 @@ pub fn main() {
     init_actions(state);
     std::env::set_var("RUST_BACKTRACE", "1");
 
-    let thread = std::thread::spawn(|| {
-        info!("hello from thread main");
-        0
-    });
+    let main_thread = std::thread::spawn(|| monitor_init());
+    main_thread.join().unwrap();
+    warn!("monitor main thread exited");
+}
 
-    let r = thread.join();
-
-    info!("==> {:?}", r);
-
-    loop {}
+fn monitor_init() {
+    info!("monitor early init completed, starting init");
 }
